+++ pd-venta/src/main/java/com/contacloud/pdventa/repository/VentaRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Optional;
 
 public interface VentaRepository extends JpaRepository<Venta, Long> {
-    List<Venta> findByClienteId(Long clienteId);
-    Optional<Venta> findTopByClienteIdAndEstadoOrderByFechaEmision(Long clienteId, String estado);
+    List<Venta> findByClienteId(Integer clienteId);
+    Optional<Venta> findTopByClienteIdAndEstadoOrderByFechaEmision(Integer clienteId, String estado);
     List<Venta> findByEstado(String estado);
 
 }
